fix: address review findings on PR #367
- connection-audit.md: correct attribution of quick wins A and B — both
  were already present on main before this PR; document now says
  'pre-existing on main' instead of 'shipped'. Section 6 heading and
  item C updated to accurately describe what this PR actually delivers.

- runner.service-announce.ts: extract isSameServiceAnnounce to its own
  small module so the unit test can import just the pure comparator
  without pulling in the full runner.ts namespace and all transitive
  server dependencies.

- runner.service-announce.test.ts: update import to use the new utility
  module instead of runner.ts.

- runner.ts: import from runner.service-announce.ts and re-export for
  backward compat; remove the inline definition.
